+++ resources/views/components/layout/header.blade.php
@@ -6,7 +6,7 @@ use DistortedFusion\BladeComponents\BladeComponents;
     '[grid-area:header]',
     'flex items-center justify-between gap-2 z-10',
 
-    'bg-[var(--sidebar)] min-h-[var(--header-height,3.5rem)]',
+    'bg-[var(--header)] min-h-[var(--header-height,3.5rem)]',
     '[:where(&)]:border-be [:where(&)]:border-[var(--border)]',
 
     'fixed top-0 inset-x-0' => $fixed,